requirements.txt
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-openai==0.2.3
langgraph==0.2.39
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1